serviciu_gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
grpcio==1.60.0
grpcio-tools==1.60.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
